+++ benches/benchmark.rs
@@ -7,7 +7,7 @@ use criterion::black_box;
 use criterion::BatchSize;
 use criterion::Criterion;
 
-use ks_curve_tracer::dut::trace::TraceWithModel;
+use ks_curve_tracer::dut::trace::{TraceWithModel, TwoTerminalGuiTrace};
 use ks_curve_tracer::dut::Device;
 use ks_curve_tracer::dut::TwoTerminalDevice;
 
@@ -22,9 +22,11 @@ fn criterion_benchmark(c: &mut Criterion) {
         "Shockley model",
         move |b, trace_name| {
             let diode_device_type = TwoTerminalDevice::Diode;
-            let trace = diode_device_type
-                .load_from_csv(format!("res/{}.csv", trace_name))
-                .expect("Can't read the test trace");
+            let trace = TwoTerminalGuiTrace::from(
+                diode_device_type
+                    .load_from_csv(format!("res/{}.csv", trace_name))
+                    .expect("Can't read the test trace"),
+            );
             b.iter_batched_ref(
                 || trace.clone(),
                 |trace| {