+++ src/dut/trace/two.rs
@@ -1,7 +1,7 @@
 use crate::backend::RawTrace;
 use crate::dut::aoi::AreaOfInterest;
 use crate::dut::csv::csv_writer_from_path;
-use crate::dut::trace::{DrawableTrace, ExportableTrace, TraceWithModel, TraceWithScatterPlot};
+use crate::dut::trace::{DrawableTrace, Trace, TraceWithModel};
 use crate::gui::{MASK_HEIGHT, MASK_WIDTH, SCATTER_PLOT_ALPHA};
 use crate::model::diode::diode_model;
 use crate::model::IVModel;
@@ -13,29 +13,45 @@ use std::f64::consts::PI;
 use std::path::Path;
 use std::sync::Arc;
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct TwoTerminalTrace {
     pub trace: RawTrace,
     pub aoi: AreaOfInterest,
+}
+
+#[derive(Clone)]
+pub struct TwoTerminalGuiTrace {
+    pub trace: TwoTerminalTrace,
     pub model: Option<Arc<dyn IVModel>>,
     scatter_plot_mask: RefCell<Option<ImageSurface>>,
 }
 
 impl TwoTerminalTrace {
     pub fn from_raw_trace(trace: RawTrace, aoi: AreaOfInterest) -> Self {
-        Self {
+        Self { trace, aoi }
+    }
+}
+
+impl From<TwoTerminalGuiTrace> for TwoTerminalTrace {
+    fn from(trace: TwoTerminalGuiTrace) -> Self {
+        trace.trace
+    }
+}
+
+impl From<TwoTerminalTrace> for TwoTerminalGuiTrace {
+    fn from(trace: TwoTerminalTrace) -> Self {
+        TwoTerminalGuiTrace {
             trace,
-            aoi,
             model: None,
             scatter_plot_mask: RefCell::new(None),
         }
     }
 }
 
-impl TraceWithModel for TwoTerminalTrace {
+impl TraceWithModel for TwoTerminalGuiTrace {
     fn fill_model(&mut self) {
         if self.model.is_none() {
-            self.model = Some(Arc::new(diode_model(&self.trace)))
+            self.model = Some(Arc::new(diode_model(&self.trace.trace)))
         }
     }
 
@@ -47,7 +63,11 @@ impl TraceWithModel for TwoTerminalTrace {
     }
 }
 
-impl ExportableTrace for TwoTerminalTrace {
+impl Trace for TwoTerminalTrace {
+    fn area_of_interest(&self) -> AreaOfInterest {
+        self.aoi
+    }
+
     fn save_as_csv(&self, path: &Path) -> Result<()> {
         let mut out = csv_writer_from_path(path)?;
 
@@ -64,8 +84,8 @@ impl ExportableTrace for TwoTerminalTrace {
     }
 }
 
-impl TraceWithScatterPlot for TwoTerminalTrace {
-    fn apply_mask(&self, cr: &Context) -> Result<()> {
+impl TwoTerminalGuiTrace {
+    pub fn apply_mask(&self, cr: &Context) -> Result<()> {
         let mut scatter_plot_mask = self.scatter_plot_mask.borrow_mut();
         if scatter_plot_mask.is_none() {
             let w = MASK_WIDTH;
@@ -82,13 +102,15 @@ impl TraceWithScatterPlot for TwoTerminalTrace {
             cr.set_operator(Operator::Over);
             cr.set_source_rgba(0.0, 0.0, 0.0, SCATTER_PLOT_ALPHA);
 
-            let v_k = f64::from(w) / (self.aoi.max_v - self.aoi.min_v);
-            let v_b = -self.aoi.min_v * v_k;
+            let aoi = self.trace.aoi;
 
-            let i_k = f64::from(h) / (self.aoi.max_i - self.aoi.min_i);
-            let i_b = -self.aoi.min_i * i_k;
+            let v_k = f64::from(w) / (aoi.max_v - aoi.min_v);
+            let v_b = -aoi.min_v * v_k;
 
-            for (v, i) in self.trace.iter() {
+            let i_k = f64::from(h) / (aoi.max_i - aoi.min_i);
+            let i_b = -aoi.min_i * i_k;
+
+            for (v, i) in self.trace.trace.iter() {
                 cr.arc(v_b + v * v_k, i_b + i * i_k, 1.0, 0.0, PI * 2.0);
                 cr.fill();
             }
@@ -103,20 +125,26 @@ impl TraceWithScatterPlot for TwoTerminalTrace {
     }
 }
 
-impl DrawableTrace for TwoTerminalTrace {
+impl Trace for TwoTerminalGuiTrace {
     fn area_of_interest(&self) -> AreaOfInterest {
-        self.aoi
+        self.trace.area_of_interest()
     }
 
+    fn save_as_csv(&self, path: &Path) -> Result<()> {
+        self.trace.save_as_csv(path)
+    }
+}
+
+impl DrawableTrace for TwoTerminalGuiTrace {
     fn draw(&self, cr: &Context, v_factor: f64, i_factor: f64, height: f64) {
         let w = f64::from(MASK_WIDTH);
         let h = f64::from(MASK_HEIGHT);
 
-        let v_k = w / (self.aoi.max_v - self.aoi.min_v);
-        let v_b = self.aoi.min_v * v_k;
+        let v_k = w / (self.trace.aoi.max_v - self.trace.aoi.min_v);
+        let v_b = self.trace.aoi.min_v * v_k;
 
-        let i_k = h / (self.aoi.max_i - self.aoi.min_i);
-        let i_b = self.aoi.min_i * i_k;
+        let i_k = h / (self.trace.aoi.max_i - self.trace.aoi.min_i);
+        let i_b = self.trace.aoi.min_i * i_k;
 
         cr.save();
         cr.set_source_rgba(0.0, 0.0, 0.0, 1.0);