+++ src/gui/widgets/bjt.rs
@@ -1,9 +1,9 @@
 use crate::dut::CurrentBiasedDeviceConfig;
 use gtk::ContainerExt;
+use gtk::LabelExt;
 use gtk::SpinButtonExt;
 use gtk::SpinButtonSignals;
 use gtk::WidgetExt;
-use gtk::LabelExt;
 use noisy_float::prelude::r64;
 use noisy_float::prelude::R64;
 use relm::{Relm, Update, Widget};