+++ src/lib.rs
@@ -3,14 +3,11 @@ extern crate lazy_static;
 #[macro_use]
 extern crate measure_time;
 #[macro_use]
-extern crate serde_derive;
-#[macro_use]
 extern crate relm;
-
 #[macro_use]
 extern crate relm_derive;
-
-use structopt;
+#[macro_use]
+extern crate serde_derive;
 
 pub mod backend;
 pub mod dut;
@@ -19,4 +16,5 @@ pub mod model;
 pub mod options;
 pub mod util;
 
-pub type Result<T> = std::result::Result<T, failure::Error>;
+pub type Error = failure::Error;
+pub type Result<T> = std::result::Result<T, Error>;