+++ src/options.rs
@@ -6,7 +6,7 @@ use structopt::StructOpt;
 
 use crate::backend::Backend;
 use crate::backend::AD2;
-use crate::dut::trace::TraceWithModel;
+use crate::dut::trace::{GuiTrace, TwoTerminalGuiTrace};
 use crate::dut::Device;
 use crate::dut::TwoTerminalDevice;
 use crate::Result;
@@ -41,13 +41,15 @@ impl Opt for CliOpt {
 }
 
 impl CliOpt {
-    pub fn trace(&self) -> Result<Box<dyn TraceWithModel>> {
+    pub fn trace(&self) -> Result<Box<dyn GuiTrace>> {
         Ok(
             match &self.device.as_ref().unwrap_or(&CliBackendOption::DWF) {
-                CliBackendOption::DWF => Box::new(TwoTerminalDevice::Diode.trace(&AD2::new()?)?),
-                CliBackendOption::Csv { file } => {
-                    Box::new(TwoTerminalDevice::Diode.load_from_csv(file.as_path())?)
-                }
+                CliBackendOption::DWF => Box::new(TwoTerminalGuiTrace::from(
+                    TwoTerminalDevice::Diode.trace(&AD2::new()?)?,
+                )),
+                CliBackendOption::Csv { file } => Box::new(TwoTerminalGuiTrace::from(
+                    TwoTerminalDevice::Diode.load_from_csv(file.as_path())?,
+                )),
             },
         )
     }