+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - ReleaseDate
+### Added
+* Spinner animation during capture, load and model fitting #24
+### Changed
+* Capturing and loading a trace is now async #24
+
 ## [0.4.1] - 2019-03-16
 ### Changed
 * Tweaked the right pane a little bit